Move internal routes to their own files + add swagger doc
