+++ api/src/routes/internal/index.ts
@@ -1,0 +1,7 @@
+import providerVersions from "./providerVersions";
+import gpu from "./gpu";
+import leasesDuration from "./leasesDuration";
+import gpuModels from "./gpuModels";
+import gpuPrices from "./gpuPrices";
+
+export default [providerVersions, gpu, leasesDuration, gpuModels, gpuPrices];
